refactor: refactor utilities, refactor imports
