+++ app/lib/matador/components/index.ts
@@ -1,9 +1,9 @@
 export * from "./dot";
-export * from "./ui/Link";
-export * from "./stat-card";
-export * from "./navbar";
+export * from "./errors";
 export * from "./job-data";
 export * from "./job-info";
 export * from "./job-result";
 export * from "./jobs-table";
-export * from "./errors/index";
+export * from "./navbar";
+export * from "./stat-card";
+
