+++ app/lib/matador/components/jobs-table/JobStatusFilterModal.tsx
@@ -1,9 +1,9 @@
 import { Button, Center, Chip, Chips, Modal, Text } from "@mantine/core";
 import { useState } from "react";
-import { getStatusColor } from "../../helpers/ui-helpers";
-import { toFirstLetterUpperCase } from "../../helpers/utils";
-import type { JobStatus } from "../../types/JobStatus";
-import { JobStatuses } from "../../types/JobStatus";
+import { getStatusColor } from "~/lib/matador/helpers/application-helpers.server";
+import { toFirstLetterUpperCase } from "~/lib/matador/helpers/utils";
+import type { JobStatus } from "~/lib/matador/types/JobStatus";
+import { JobStatuses } from "~/lib/matador/types/JobStatus";
 
 export interface JobStatusFilterModalProps {
   opened: boolean;
