--- app/lib/matador/components/ui/Link.tsx
@@ -1,24 +1,0 @@
-import { Link as RemixLink } from "@remix-run/react";
-
-const IS_EXTERNAL_LINK_REGEX = /^(?:[a-z][a-z\d+.-]*:|\/\/)/;
-
-export function Link({ children, url = "", external, ref, ...rest }: any) {
-  // react-router only supports links to pages it can handle itself. It does not
-  // support arbirary links, so anything that is not a path-based link should
-  // use a reglar old `a` tag
-  if (external || IS_EXTERNAL_LINK_REGEX.test(url)) {
-    rest.target = "_blank";
-    rest.rel = "noopener noreferrer";
-    return (
-      <a href={url} {...rest}>
-        {children}
-      </a>
-    );
-  }
-
-  return (
-    <RemixLink to={url} {...rest}>
-      {children}
-    </RemixLink>
-  );
-}
