--- app/lib/matador/constants.ts
@@ -1,1 +1,0 @@
-export const themeKeyLocalStorage = "matador_theme";
