--- app/lib/matador/helpers/style-helpers.ts
@@ -1,3 +1,0 @@
-export function classNames(...classes: any[]): string {
-  return classes.filter(Boolean).join(" ");
-}
